Adressed in last commit
